Allow RefreshTokenGrant to validate client / token relationship
Servers may bind tokens to client IDs to reduce chance of token abuse, but `RefreshTokenGrant` does not allow for this functionality.

In `validate_token_request`, the client is validated before the user function `authenticate_refresh_token` is called, but the client isn't saved as an instance variable until after that function call.  This change allows `authenticate_refresh_token` to access the client with `self.request.client`