OAuth2Client: use correct auth method for token introspection
When token introspection was introduced in 6f5d19a, using the
code that previously only handled token revocation, the new
`_handle_token_hint` method that does the work for both
`introspect_token` and `revoke_token` kept using
`self.revocation_endpoint_auth_method` unconditionally if no
`auth` was passed in with the introspect or revoke request.
This seems to be wrong, introspecting a token should use the
`token_endpoint_auth_method`.

This leaves the fallback to `revocation_endpoint_auth_method`
in `_handle_token_hint` because adjusting its signature to make
`auth` compulsory would be awkward, but it's not expected ever
to be used.

Signed-off-by: Adam Williamson <awilliam@redhat.com>